refactor: merge conversation message renderers into base renderers
Eliminates the wrapper pattern by merging todo list support and socket
real-time updates directly into the base renderers:

- MessageRenderer now handles both standard and todo_list messages
- CompactMessageRenderer now handles both standard and todo_list messages
- Delete ConversationMessageRenderer.tsx (was wrapper around MessageRenderer)
- Delete CompactConversationMessageRenderer.tsx (was wrapper around CompactMessageRenderer)
- Update ChatMessagesArea to use MessageRenderer directly (remove toggle)
- Update AssistantChatTab to use CompactMessageRenderer directly

This simplifies the architecture by having a single renderer per size
variant (full-size and compact) instead of base + wrapper pattern.
